+++ konlpy/tag/_mecab.py
@@ -55,15 +55,25 @@ class Mecab():
     .. _Eunjeon Project: http://eunjeon.blogspot.kr/
     """
 
-    def pos(self, phrase):
+    # TODO: check whether flattened results equal non-flattened
+    def pos(self, phrase, flatten=True):
         """POS tagger."""
         if sys.version_info[0] < 3:
             phrase = utils.preprocess(phrase).encode('utf-8')
-            result = self.tagger.parse(phrase).decode('utf-8')
+            if flatten:
+                result = self.tagger.parse(phrase).decode('utf-8')
+                return parse(result)
+            else:
+                return [parse(self.tagger.parse(eojeol).decode('utf-8'))\
+                        for eojeol in phrase.split()]
         else:
             phrase = utils.preprocess(phrase)
-            result = self.tagger.parse(phrase)
-        return parse(result)
+            if flatten:
+                result = self.tagger.parse(phrase)
+                return parse(result)
+            else:
+                return [parse(self.tagger.parse(eojeol).decode('utf-8'))\
+                        for eojeol in phrase.split()]
 
     def morphs(self, phrase):
         """Parse phrase to morphemes."""